backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.2.0
python-multipart==0.0.6